+++ checkout-ui-custom/src/controller/AddressController.js
@@ -1,4 +1,6 @@
 import intlTelInput from 'intl-tel-input';
+// the intlTelInput library has dependencies with ...js/utils import, do not remove it if using the intlTelInput library
+import 'intl-tel-input/build/js/utils';
 import 'intl-tel-input/build/css/intlTelInput.css';
 
 import {
@@ -7,6 +9,7 @@ import {
   STEPS,
   TIMEOUT_500
 } from '../utils/const';
+import setTranslations from '../utils/translations';
 
 const AddressController = (() => {
   const state = {
@@ -29,7 +32,8 @@ const AddressController = (() => {
       const iti = intlTelInput(phoneInput, {
         initialCountry: COUNTRIES.za.code,
         onlyCountries: COUNTRIES_AVAILABLES,
-        customPlaceholder
+        customPlaceholder,
+        formatOnDisplay: false
       });
       state.intTelInput = iti;
 
@@ -39,6 +43,74 @@ const AddressController = (() => {
       }
 
       phoneInput.focus();
+    }
+  };
+
+  const setAddressType = () => {
+    const addressTypeInput = document.querySelector('.vtex-omnishipping-1-x-address .ship-addressType');
+
+    if (!addressTypeInput) {
+      $('.vtex-omnishipping-1-x-address > div').append(
+        $('<p>').prop({
+          class: 'input ship-addressType text',
+          style: 'order: 3'
+        })
+      );
+
+      $('.ship-addressType').append(
+        $('<label>').html('Address type')
+      );
+
+      $('.ship-addressType').append(
+        $('<div>').prop({
+          class: 'ship-addressType-container'
+        })
+      );
+
+      $('.ship-addressType-container').append(
+        $('<div>').prop({
+          class: 'ship-addressType-div-residential'
+        })
+      );
+
+      $('.ship-addressType-container').append(
+        $('<div>').prop({
+          class: 'ship-addressType-div-business'
+        })
+      );
+
+      $('.ship-addressType-div-residential').append(
+        $('<input>').prop({
+          type: 'radio',
+          id: 'ship-addressType-residential',
+          name: 'ship-addressType',
+          value: 'residential'
+        })
+      ).append(
+        $('<label>').prop({
+          for: 'ship-addressType-residential'
+        }).html('Residential')
+      );
+
+      $('.ship-addressType-div-business').append(
+        $('<input>').prop({
+          type: 'radio',
+          id: 'ship-addressType-business',
+          name: 'ship-addressType',
+          value: 'commercial'
+        })
+      ).append(
+        $('<label>').prop({
+          for: 'ship-addressType-business'
+        }).html('Business')
+      );
+
+      const addressTypeSelected = window.vtexjs.checkout.orderForm.shippingData?.address?.addressType;
+      if (addressTypeSelected === 'residential') {
+        $('#ship-addressType-residential').attr('checked', true);
+      } else if (addressTypeSelected === 'commercial') {
+        $('#ship-addressType-business').attr('checked', true);
+      }
     }
   };
 
@@ -67,9 +139,10 @@ const AddressController = (() => {
     if (window.location.hash === STEPS.SHIPPING) {
       setTimeout(() => {
         const selectedDelivery = $('#shipping-option-delivery').hasClass('shp-method-option-active');
-
+        setTranslations();
         if (window.location.hash === STEPS.SHIPPING && selectedDelivery) {
           setInputPhone();
+          setAddressType();
           toggleGoogleInput();
         }
       }, TIMEOUT_500);
@@ -93,6 +166,10 @@ const AddressController = (() => {
           .attr('value', window.vtexjs.checkout.orderForm.shippingData.address.complement || '');
       }
     }, TIMEOUT_500);
+  });
+
+  $(document).on('change', 'input[name="ship-addressType"]', () => {
+    localStorage.setItem('addressType', document.querySelector('input[name="ship-addressType"]:checked').value);
   });
 
   const publicInit = () => { };