+++ checkout-ui-custom/src/utils/translations.js
@@ -1,0 +1,3 @@
+export default function setTranslations() {
+  $('.btn-go-to-shippping-method').text('Save address');
+}